+++ tests/testthat/test-expect_filename_error.R
@@ -1,5 +1,4 @@
 test_that("Survey exists to read from disk", {
-  # Call getsurvey
-  expect_error(qualtRics::readSurvey("/users/jasper/desktop/error.csv"),
+  expect_error(qualtRics::readSurvey("/users/julia/desktop/error.csv"),
                "does not exist")
 })