+++ tests/testthat/test-getSurvey.R
@@ -5,7 +5,7 @@ with_mock_api({
   test_that("getSurvey() returns a survey", {
     testthat::skip_on_cran()
     Sys.setenv("QUALTRICS_WARNING_DATE_GIVEN"=TRUE)
-    registerOptions(base_url="t.qualtrics.com", api_token="1234")
+    qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
     # Get survey
     surveys <- getSurvey("SV_5u9zu8zHnHaGml7")
     # TESTS