+++ tests/testthat/test-root_url_fails.R
@@ -1,7 +1,7 @@
-test_that("Registering root url that doesn't end with '.qualtrics.com' fails.", { # nolint
+test_that("using getSurvey() with a base URL that doesn't end with '.qualtrics.com' fails", { # nolint
   # Store dummy key
-  qualtRics::registerOptions(api_token="1234", base_url="abcd")
+  qualtrics_api_credentials(api_key = "1234", base_url = "abcd")
   # This should error on checkParams
   expect_error(qualtRics::getSurveys(),
-               "The qualtrics root url must end with '.qualtrics.com'. Your root url looks like this: 'abcd'.\nPlease visit https://api.qualtrics.com/docs/root-url for instructions about the qualtrics root url.") # nolint
+               "The Qualtrics base URL must end with") # nolint
 })